requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
httpx==0.25.2
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0
python-multipart==0.0.6
aiofiles==23.2.1